started working on functions to turn command numbers to timer delays

+++ inc/IR_lib.h
@@ -7,6 +7,6 @@
 #include "cmd_assoc_structs.h"
 
 void send_command(void);
-int16_t execute_command(const struct command *cmd, bool is_ditto);
+int32_t execute_command(const struct command *cmd, bool is_ditto);
 
 #endif /* __IR_LIB_H */
